Move execution to .py
