+++ .gitignore
@@ -110,9 +110,7 @@ dmypy.json
 # End of https://www.gitignore.io/api/python
 
 .ipynb_checkpoints/
-
-
+*.ipynb
 
 # mac
-
 .DS_Store
